fix #3

+++ src/main/java/vavi/util/box/Meta.java
@@ -55,7 +55,7 @@ public interface Meta {
                 Box subBox = factory.getInstance(dis);
 //System.err.println(subBox);
                 subBoxes.add(subBox);
-                l += subBox.offset;
+                l += (int) subBox.offset;
 //Debug.println(l + "/" + (offset - 8));
             }
         }
